correcao da variavel de inteiro

+++ src/PersonDataStorage.sol
@@ -3,6 +3,6 @@ pragma solidity ^0.8.13;
 
 contract Counter {
     string public name = "Marco Antonio";
-    uint private age = 32;
+    uint256 private age = 32;
     bool internal brazilian = true;
 }
